Move factories to the same file as the views they produce
