+++ android/src/main/kotlin/com/airship/flutter/FlutterEmbeddedView.kt
@@ -10,6 +10,9 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.platform.PlatformView
 import com.urbanairship.embedded.AirshipEmbeddedView
+import io.flutter.plugin.common.BinaryMessenger
+import io.flutter.plugin.platform.PlatformView
+import io.flutter.plugin.platform.PlatformViewFactory
 
 class FlutterEmbeddedView(
     private var context: Context,
@@ -23,5 +26,19 @@ class FlutterEmbeddedView(
     override fun dispose() {}
     override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
         // Handle method calls if needed
+    }
+}
+
+
+class EmbeddedViewFactory(private val binaryMessenger: BinaryMessenger) : PlatformViewFactory(StandardMessageCodec.INSTANCE) {
+    override fun create(context: Context?, viewId: Int, args: Any?): PlatformView {
+        val channel = MethodChannel(binaryMessenger, "com.airship.flutter/EmbeddedView_$viewId")
+
+        // Extracting embeddedId from args
+        val params = args as? Map<String, Any>
+        val embeddedId = params?.get("embeddedId") as? String ?: "defaultId"
+
+        val view = FlutterEmbeddedView(checkNotNull(context), channel, embeddedId)
+        return view
     }
 }
