--- android/src/main/kotlin/com/airship/flutter/InboxMessageViewFactory.kt
@@ -1,17 +1,0 @@
-package com.airship.flutter
-
-import android.content.Context
-import io.flutter.plugin.common.BinaryMessenger
-import io.flutter.plugin.common.MethodChannel
-import io.flutter.plugin.common.StandardMessageCodec
-import io.flutter.plugin.platform.PlatformView
-import io.flutter.plugin.platform.PlatformViewFactory
-
-class InboxMessageViewFactory(private val binaryMessenger: BinaryMessenger) : PlatformViewFactory(StandardMessageCodec.INSTANCE) {
-    override fun create(context: Context?, viewId: Int, args: Any?): PlatformView {
-        val channel = MethodChannel(binaryMessenger, "com.airship.flutter/InboxMessageView_$viewId")
-        val view = FlutterInboxMessageView(checkNotNull(context), channel)
-        channel.setMethodCallHandler(view)
-        return view
-    }
-}
